Set paver command tests to run as part of unit test lib suites.
